update: restructure folders, remap containers
