+++ docker-compose.override.yml
@@ -1,22 +1,26 @@
 services:
   gateway:
     volumes:
-      - ./gateway/src:/app/src
+      - ./gateway/:/app
+      - /app/node_modules
     command: npm run dev
 
   auth-service:
     volumes:
-      - ./services/auth-service/src:/app/src
+      - ./services/auth-service/:/app
+      - /app/node_modules
     command: npm run dev
 
   quiz-service:
     volumes:
-      - ./services/quiz-service/src:/app/src
+      - ./services/quiz-service/:/app
+      - /app/node_modules
     command: npm run dev
 
   result-service:
     volumes:
-      - ./services/result-service/src:/app/src
+      - ./services/result-service/:/app
+      - /app/node_modules
     command: npm run dev
   frontend:
       build:
