+++ services/quiz-service/utils/errorHandler.js
